+++ src/main/java/io/fno/grel/OtherFunctions.java
@@ -10,8 +10,16 @@ import java.lang.reflect.Field;
 public class OtherFunctions {
 
     // https://docs.openrefine.org/manual/grelfunctions#typeo
-    public static Class<?> type(Object o) {
-        return o.getClass();
+
+    /**
+     * <a href="https://docs.openrefine.org/manual/grelfunctions#typeo">type</a>
+     * <br>
+     * Returns a string with the data type of o, such as undefined, string, number, boolean, etc.
+     * @param o The object to get the type of.
+     * @return  The simple name of the class of {@code o} in lower case.
+     */
+    public static String type(Object o) {
+        return o.getClass().getSimpleName().toLowerCase();
     }
 
     // https://docs.openrefine.org/manual/grelfunctions#hasfieldo-s-name